changes to repo

+++ archer_scripts/opt_script.csh
@@ -5,7 +5,7 @@
 #
 
 #nprocs=128
-itermax=20
+itermax=25
 procsonnode=128
 
 
@@ -22,6 +22,9 @@ echo $ite
 i=`expr $ite + 1`
 echo "HERE2"
 echo $itermax
+
+mkdir gradcontrol
+
 while [ $i -le $itermax ]
 do
  echo "GOT HERE"
@@ -33,13 +36,10 @@ do
  mv -f TTT.tmp data.optim
  fich=output$name$ii
  echo "Running mitcgm_ad: iteration $ii"
- ls mitgcmuv_ad
-# mpirun -n $nprocs ./new.csh
-# srun -n $nprocs -N $procsonnode ./mitgcmuv_ad
  srun --distribution=block:block --hint=nomultithread ./mitgcmuv_ad > out 2> err
  rm tapelev*
  rm oad_cp*
- cp STDOUT.0000 $fich
+ mv STDOUT.0000 $fich
  egrep optimcycle data.optim >> fcost$name
  grep "objf_temp_tut(" $fich >> fcost$name
  grep "objf_hflux_tut(" $fich >> fcost$name
@@ -47,11 +47,16 @@ do
  grep 'global fc =' $fich
  echo Cleaning
  \rm tapelev*
- direc=run$name$ii
- mkdir $direc
- rm pickup.* pickup_*  maskCtrl* hFac* wunit* RA*ta DX*ta DY*ta DR*ta PH*ta
- mv -f *.meta *.data STDOUT* STDERR* $direc
- mv -f $direc/wunit*.*data ./
+ cp adxx* gradcontrol
+ cp xx* gradcontrol
+ if [ `expr $i % 5` -eq 0 ]
+ then
+  direc=run$name$ii
+  mkdir $direc
+  rm pickup.* pickup_*  maskCtrl* hFac* wunit* RA*ta DX*ta DY*ta DR*ta PH*ta
+  mv -f *.meta *.data STDOUT* STDERR* out err $direc 
+  mv -f $direc/wunit*.*data ./
+ fi
  cp -f ecco_ctrl_MIT_CE_000.opt0$ii OPTIM/
  cp -f ecco_cost_MIT_CE_000.opt0$ii OPTIM/
  echo "Line-search: iteration $ii"
@@ -66,12 +71,14 @@ do
  echo $i
 done
 
+rm tapelev*
+rm oad_cp*
+
 exit
 
 for i in $(ls -d runoptiter*00); do 
  mv $i SAVE$i;
 done
-rm -r runoptiter*
 rm OPTIM/OPWARM*
